extracted the logout button into its own component
